Task highlighting introduced.
Tasks are now highlighted based on their associated date: tasks to be done today
or before are in red, all others in green. Currently loads the date list from the
preferences file twice, which needs sorting out, and also doesn't update the
highlighting  when a date is changed until the app is restarted.

Version number updated in manifest and versions plan updated.