Multiple list functionality added
Very basic ability to use multiple lists, implemented in exactly the same way as
the tasks on an individual list. Currently list names are not editable once the
list is added, as clicking a list leads to the tasks on that list.

Contains a great big buggy bug which causes the app to crash if opened from the
unlock screen rather than from the apps menu. Versions plan has been updated to
accomodate this in the next bugfix release.